+++ MyLayout/Lib/MyLayoutInner.h
@@ -210,8 +210,14 @@ extern BOOL _myCGFloatEqual(CGFloat f1, CGFloat f2);
 extern BOOL _myCGFloatNotEqual(CGFloat f1, CGFloat f2);
 extern BOOL _myCGFloatLessOrEqual(CGFloat f1, CGFloat f2);
 extern BOOL _myCGFloatGreatOrEqual(CGFloat f1, CGFloat f2);
+extern BOOL _myCGSizeEqual(CGSize sz1, CGSize sz2);
+extern BOOL _myCGPointEqual(CGPoint pt1, CGPoint pt2);
+extern BOOL _myCGRectEqual(CGRect rect1, CGRect rect2);
+
+
 extern CGFloat _myRoundNumber(CGFloat f);
 extern CGRect _myRoundRect(CGRect rect);
 extern CGSize _myRoundSize(CGSize size);
 extern CGPoint _myRoundPoint(CGPoint point);
+extern CGRect _myRoundRectForLayout(CGRect rect);