+++ MyLayoutDemo/AllTest2TableViewCell.m
@@ -32,8 +32,8 @@
          * 您可以尝试用不同的布局来实现相同的功能。
          */
         //[self createLinearRootLayout];
-       // [self createRelativeRootLayout];
-         [self createFloatRootLayout];
+        [self createRelativeRootLayout];
+       //  [self createFloatRootLayout];
     }
     
     return self;